Adicionando rotas e validacao de servicos.
