--- backend/routes.js
@@ -1,20 +1,0 @@
-const { Router } = require('express');
-const UserController = require('./src/controllers/UserController');
-const ServiceController = require('./src/controllers/ServiceController');
-
-const routes = Router();
-
-routes.get('/', UserController.firstMessage);
-routes.get('/user/:id', UserController.read);
-routes.post('/user', UserController.create);
-routes.put('/user/:id', UserController.update);
-routes.delete('/user/:id', UserController.delete);
-
-routes.post('/service', ServiceController.create);
-routes.get('/service/:id', ServiceController.read);
-routes.put('/service/:id', ServiceController.update);
-routes.delete('/service/:id', ServiceController.delete);
-
-routes.get('/teste',ServiceController.Message);
-
-module.exports = routes;
